[#26] feature: 조건에 따른 총 가격 기능 추가
- Guest, Price, ReservationDate는 사용하지 않아서 삭제
- RoomMapper
  - totalPrice(): 슈퍼호스트 존재여부에 따른 총가격 표시

+++ BE/src/main/java/com/codesquad/airbnb/utils/DayCalculator.java
@@ -5,9 +5,9 @@ import java.util.concurrent.TimeUnit;
 
 public class DayCalculator {
 
-    public static long getDiffDays(Date day1, Date day2) {
+    public static int getDiffDays(Date day1, Date day2) {
         long diff = day2.getTime() - day1.getTime();
         TimeUnit time = TimeUnit.MILLISECONDS;
-        return time.toDays(diff);
+        return (int) time.toDays(diff);
     }
 }
